Added OS detection to bash configuration files

+++ .bashrc
@@ -5,36 +5,29 @@
 # If not running interactively, don't do anything
 [[ $- != *i* ]] && return
 
-# Show and hide hidden files
-alias showFiles='defaults write com.apple.finder AppleShowAllFiles YES;
-killall Finder /System/Library/CoreServices/Finder.app'
-alias hideFiles='defaults write com.apple.finder AppleShowAllFiles NO;
-killall Finder /System/Library/CoreServices/Finder.app'
+# OS Specific
+if [[ "$(uname)" == 'Darwin' ]]; then
+  # Show and hide hidden files
+  alias showFiles='defaults write com.apple.finder AppleShowAllFiles YES;
+  killall Finder /System/Library/CoreServices/Finder.app'
+  alias hideFiles='defaults write com.apple.finder AppleShowAllFiles NO;
+  killall Finder /System/Library/CoreServices/Finder.app'
 
-# Reset the screen shot directory to the desktop
-alias resetScreenshotdir='defaults write com.apple.screencapture location ~/Desktop; killall SystemUIServer'
+  alias py='python3'
+
+  # Pretty code copy function
+  function hl() {
+      if [ -z "$1" ]; then
+          echo Your doing things wrong...
+      else
+          highlight -O rtf -t 2 -K 11 "$1" | pbcopy
+      fi
+  }
+elif [[ "$(uname)" == 'Linux' ]]; then
+  alias sysupdate ='sudo pacman -Syu'
+fi
 
 alias cl='clear'
 alias ll='ls -lah'
 alias tree='tree -I *.pyc'
 
-# Pretty code copy function
-function hl() {
-    if [ -z "$1" ]; then
-        echo Your doing things wrong...
-    else
-        highlight -O rtf -t 2 -K 11 "$1" | pbcopy
-    fi
-}
-
-# Add laravel to the PATH. Gets removed on close, hence here.
-export PATH="$PATH:~/.composer/vendor/bin"
-
-# Add homebrew sbin to PATH so that executables work
-export PATH="$PATH:/usr/local/sbin"
-
-# Add Mercurial to PATH
-export PATH="$PATH:/usr/local/Cellar/mercurial/3.5.2/bin"
-
-# Set the EDITOR env variable
-export EDITOR=vim
